SDLC Linux VM Instructions
How to connect to our VM and run the version of the app previously setup on the SDLC Ubuntu VM. Note: does not cover setting up new versions of the app, for that see the "readme" file. You can either clone an existing copy already on the VM or download a fresh one from GitHub and follow other setup instructions to have your own VM environment.